% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_correct.R
\name{compute_prob_correct}
\alias{compute_prob_correct}
\title{Compute each biomarker's probability of being recorded at its baseline level}
\usage{
compute_prob_correct(dataset, biomarker_levels, max_prob = 1)
}
\arguments{
\item{dataset}{a \code{\link[=data.frame]{data.frame()}}; should be just control data}

\item{biomarker_levels}{a list containing the levels of each biomarker}

\item{max_prob}{a maximum value for the baseline probability}
}
\value{
a \code{\link[=structure]{structure()}} that is fundamentally a named \code{\link[=numeric]{numeric()}} vector
corresponding to the elements of \code{biomarker_levels},
but also contains a \link[tibble:tbl_df]{tibble::tbl_df} as attribute \code{data}
(easier than fixing all uses of this function)
}
\description{
Compute each biomarker's probability of being recorded at its baseline level
}
\examples{
require(dplyr)
full_data <- sim_data
biomarker_group_list = list(
    "group 1" = c("Biomarker 1", "Biomarker 2"),
   "group 2" = c("Biomarker 3", "Biomarker 4"),
  "group 3" = "Biomarker 5"
  )
biomarker_groups <- compile_biomarker_groups_table(
dataset = full_data,
biomarker_group_list = biomarker_group_list)

biomarker_varnames <-
  biomarker_groups |>
  dplyr::pull("biomarker")

biomarker_levels <-
  full_data |>
  get_levels(biomarker_varnames)

control_data <-
  full_data |>
  dplyr::filter(Category == "Patient") |>
  dplyr::select(all_of(biomarker_varnames))

control_data |>
  compute_prob_correct(
    max_prob = .95,
    biomarker_levels = biomarker_levels
  ) |>
  attr("data")

}
