% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Actions taken when \code{rSuStaIn} is loaded}
\usage{
.onLoad(...)
}
\arguments{
\item{...}{not used}
}
\value{
\code{\link{NULL}}, invisibly
}
\description{
based on https://rstudio.github.io/reticulate/articles/package.html
#delay-loading-python-modules
}
