% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results_from_pickle.R
\name{extract_results_from_pickle}
\alias{extract_results_from_pickle}
\title{Extract results from pickle file}
\usage{
extract_results_from_pickle(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  rda_filename = "data.RData",
  basename = paste0(dataset_name, "_subtype", n_s - 1),
  picklename = paste0(basename, ".pickle"),
  format_results = FALSE,
  subtype_order = seq_len(n_s),
  use_rds = TRUE,
  save_rds = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_s}{number of latent subgroups; helps construct \code{picklename}}

\item{dataset_name}{root name of dataset}

\item{output_folder}{where to find the dataset}

\item{rda_filename}{name of rda file containing environment used to run analyses}

\item{basename}{base of pickle file name (without \code{.pickle} suffix)}

\item{picklename}{the name of the pickle file to open}

\item{format_results}{whether to format results before returning}

\item{subtype_order}{an \link{integer} vector of length \code{n_s},
indicating how to order the subtypes}

\item{use_rds}{\link{logical} whether to use previously cached results}

\item{save_rds}{a \link{logical} indicating whether to save new cached results}

\item{verbose}{whether to print messages}

\item{...}{additional arguments (currently unused)}
}
\value{
a list containing the extracted results
}
\description{
Extract results from pickle file
}
\examples{
\dontrun{
output_path <-
  fs::path_package("extdata/sim_data", package = "rSuStaIn")
results <- extract_results_from_pickle(
  n = 3,
  output_folder = output_path, use_rds = FALSE
)
}
}
