% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_levels.R
\name{get_levels}
\alias{get_levels}
\title{Extract levels from a set of variable names}
\usage{
get_levels(data, varnames = names(data), keep_labels = TRUE)
}
\arguments{
\item{data}{a \link{data.frame}}

\item{varnames}{a \link{character} specifying which columns of \code{data} to analyze}

\item{keep_labels}{\link{logical} whether to preserve variable labels from \code{data} in output}
}
\value{
a \link{list} of \link{character} vectors containing
the levels of the columns in \code{data} specified by \code{varnames}
}
\description{
Extract levels from a set of variable names
}
\examples{
get_levels(iris, "Species")
}
