% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_correct.R
\name{pander.prob_correct}
\alias{pander.prob_correct}
\title{Print \code{prob_correct} objects as Pandoc markdown tables}
\usage{
\method{pander}{prob_correct}(x, ...)
}
\arguments{
\item{x}{a \code{prob_correct} object (from \code{\link[=compute_prob_correct]{compute_prob_correct()}})}

\item{...}{currently unused}
}
\value{
By default this function outputs (see: \code{cat}) the result. If you would want to catch the result instead, then call the function ending in \code{.return}.
}
\description{
This function is a method of the \code{\link[pander:pander]{pander::pander()}} function
from the  \code{pander} package, for \code{prob_correct} objects
(from \code{\link[=compute_prob_correct]{compute_prob_correct()}}).
It prints \code{prob_correct} objects as Pandoc markdown tables.
}
\examples{
full_data <- test_data_v1
v1_usable <- full_data |> dplyr::filter(CGG < 200)

biomarker_groups <- compile_biomarker_groups_table(
   dataset = v1_usable)

biomarker_varnames <-
  biomarker_groups |>
  dplyr::pull("biomarker")

biomarker_levels <-
  v1_usable |>
  get_levels(biomarker_varnames)

control_data <-
  v1_usable |>
  dplyr::filter(`FX*` == "CGG <55") |>
  dplyr::select(all_of(biomarker_varnames))

control_data |>
  compute_prob_correct(
    max_prob = .95,
    biomarker_levels = biomarker_levels
  ) |>
  attr("data") |>
  pander::pander()

}
