% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSuStaIn-package.R
\docType{package}
\name{rSuStaIn-package}
\alias{rSuStaIn}
\alias{rSuStaIn-package}
\title{rSuStaIn: R Wrapper for Python 'pySuStaIn' Package}
\description{
Provides an R interface to the 'pySuStaIn' Python package (\url{https://github.com/ucl-pond/pySuStaIn}) for Subtype and Stage Inference (SuStaIn). SuStaIn is an unsupervised learning algorithm that identifies disease subtypes and stages from cross-sectional data. This package includes wrapper functions for the main SuStaIn algorithms: Ordinal SuStaIn for ordinal/scored data, Z-score SuStaIn for continuous biomarkers, and Mixture Model SuStaIn for event-based modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ucdavis/rSuStaIn}
  \item \url{https://ucdavis.github.io/rSuStaIn/}
  \item Report bugs at \url{https://github.com/ucdavis/rSuStaIn/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu}

Authors:
\itemize{
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu}
  \item Matthew Ponzini \email{mdponzini@ucdavis.edu}
}

}
\keyword{internal}
