% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 10 columns.
}
\usage{
sim_data
}
\description{
Toy example data provided by Alexandra Young and modified by adding variables
\code{Sex} and \code{Age}
}
\keyword{datasets}
