% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_subtype_and_stage_table}
\alias{sim_subtype_and_stage_table}
\title{Simulated subtype-and-stage table}
\format{
\subsection{\code{sim_subtype_and_stage_table}}{

A data frame with nrow rows and ncol columns:
\describe{
\item{col1}{col1name}
}
}
}
\usage{
sim_subtype_and_stage_table
}
\description{
See \code{vignettes/articles/SuStaIn-simulated-data.qmd} and
\code{data-raw/sim_subtype_and_stage_table.R}
for details on how this file was created.
}
\keyword{datasets}
