% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_barplot.R
\name{stage_barplot.default}
\alias{stage_barplot.default}
\title{Plot estimated stage counts}
\usage{
\method{stage_barplot}{default}(
  object,
  include_type_0 = TRUE,
  xmax = max(object$ml_stage, na.rm = TRUE),
  ...
)
}
\arguments{
\item{object}{a \code{subtype_and_stage_table} object
(a type of data.frame)}

\item{include_type_0}{\link{logical} whether to include Type 0 in the plot}

\item{xmax}{largest x-value to display}

\item{...}{unused}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot estimated stage counts
}
\examples{
sim_subtype_and_stage_table |> stage_barplot()
}
