% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_barplot.R
\name{stage_barplot.list}
\alias{stage_barplot.list}
\title{Plot estimated stage counts}
\usage{
\method{stage_barplot}{list}(object, ...)
}
\arguments{
\item{object}{a \link{list} of \code{Sustain_model} objects}

\item{...}{
  Arguments passed on to \code{\link{stage_barplot.default}}
  \describe{
    \item{\code{xmax}}{largest x-value to display}
    \item{\code{include_type_0}}{\link{logical} whether to include Type 0 in the plot}
  }}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot estimated stage counts
}
