% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_barplot.SuStaIn_model.R
\name{stage_barplot.SuStaIn_model}
\alias{stage_barplot.SuStaIn_model}
\title{Plot estimated stage counts}
\usage{
\method{stage_barplot}{SuStaIn_model}(object, ...)
}
\arguments{
\item{object}{a \code{SuStaIn_model} object}

\item{...}{
  Arguments passed on to \code{\link{stage_barplot.default}}
  \describe{
    \item{\code{xmax}}{largest x-value to display}
    \item{\code{include_type_0}}{\link{logical} whether to include Type 0 in the plot}
  }}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot estimated stage counts
}
\examples{
output_path <-
  fs::path_package("extdata/sim_data", package = "rSuStaIn")

results <-
  extract_results_from_pickle(
    output_folder = output_path,
    use_rds = TRUE,
    n_s = 3
  )

results |>
  stage_barplot()
}
