% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data}
\format{
\subsection{\code{test_data}}{

A data frame with nrow rows and ncol columns:
\describe{
\item{col1}{col1name}
}
}
}
\usage{
test_data
}
\description{
Scrambled data for testing
}
\keyword{datasets}
